Update README.md
India is an agrarian economy, that is, its major revenue comes
from the fields. Which is why it is essential to come up with
newer technologies which save resources and money. A major
problem is irrigation. The irrigation process is manual, where
farmhands turn on/off the irrigation system when needed. This
can be an issue, because watering is done purely on speculation
and sometimes the fields could be overwatered.
Our solution is to build a smart irrigation system. It detects the
moisture content of the soil based on which it turns the system
on/off.
This works on the principle that when the soil is wet, electricity
will start flowing through it and thus the potential drop across
two points (sensing terminals) will decrease
When the soil is dry, there is no electricity flowing through the
soil and the potential drop across two points will increase.
We will be using this increase and decrease in the potential drop
to operate our system.